:goal_net: Check if installer script exists

+++ templates/docker/prestashop/.docker/php-fpm/run.sh
@@ -67,16 +67,20 @@ if [ ! -f /var/www/app/config/settings.inc.php ]; then
             export PS_DOMAIN=$(hostname -i)
         fi
 
-        echo "* Launching the installer script..."
-        su - www-data -s /bin/bash -c "/usr/local/bin/php /var/www/app/$PS_FOLDER_INSTALL/index_cli.php \
-        --domain=\"$PS_DOMAIN\" --db_server=$DB_SERVER:$DB_PORT --db_name=\"$DB_NAME\" --db_user=$DB_USER \
-        --db_password=$DB_PASSWD --prefix=\"$DB_PREFIX\" --firstname=\"$ADMIN_FIRSTNAME\" --lastname=\"$ADMIN_LASTNAME\" \
-        --password=$ADMIN_PASSWD --email=\"$ADMIN_MAIL\" --language=$PS_LANGUAGE --country=$PS_COUNTRY \
-        --all_languages=$PS_ALL_LANGUAGES --newsletter=0 --send_email=0 --ssl=$PS_ENABLE_SSL"
+        if [ -f /var/www/app/$PS_FOLDER_INSTALL/index_cli.php ]; then
+            echo "* Launching the installer script..."
+            su - www-data -s /bin/bash -c "/usr/local/bin/php /var/www/app/$PS_FOLDER_INSTALL/index_cli.php \
+            --domain=\"$PS_DOMAIN\" --db_server=$DB_SERVER:$DB_PORT --db_name=\"$DB_NAME\" --db_user=$DB_USER \
+            --db_password=$DB_PASSWD --prefix=\"$DB_PREFIX\" --firstname=\"$ADMIN_FIRSTNAME\" --lastname=\"$ADMIN_LASTNAME\" \
+            --password=$ADMIN_PASSWD --email=\"$ADMIN_MAIL\" --language=$PS_LANGUAGE --country=$PS_COUNTRY \
+            --all_languages=$PS_ALL_LANGUAGES --newsletter=0 --send_email=0 --ssl=$PS_ENABLE_SSL"
 
-        mv /var/www/app/admin /var/www/app/$PS_FOLDER_ADMIN
-        rm -rf /var/www/app/PS_FOLDER_INSTALL
-        rm /var/www/app/prestashop.zip
+            mv /var/www/app/admin /var/www/app/$PS_FOLDER_ADMIN
+            rm -rf /var/www/app/PS_FOLDER_INSTALL
+            rm /var/www/app/prestashop.zip
+        else
+            echo "Installer script is missing, cannot install"
+        fi
 
         if [ $? -ne 0 ]; then
             echo 'warning: PrestaShop installation failed.'
